Rename atoms common
