+++ src/components/DepositWidget/Form.tsx
@@ -9,7 +9,7 @@ import { formatAmountFull, parseAmount, abbreviateString } from 'utils'
 
 // Components
 import { CardDrawer } from 'components/Layout/Card'
-import { Spinner } from 'components/atoms/Spinner'
+import { Spinner } from 'components/common/Spinner'
 
 // DepositWidget: subcomponents
 import { WalletDrawerInnerWrapper } from 'components/DepositWidget/Form.styled'
