+++ src/components/PoolingWidget/DefineSpread.tsx
@@ -7,7 +7,7 @@ import { HelpTooltipContainer, HelpTooltip } from 'components/Tooltip'
 import { Input } from 'components/Input'
 
 // TradeWidget: subcomponent
-import { FormInputError } from 'components/atoms/FormInputError'
+import { FormInputError } from 'components/common/FormInputError'
 
 import { useFormContext, FieldError } from 'react-hook-form'
