+++ src/components/TradeWidget/Price.tsx
@@ -13,8 +13,8 @@ import { OrderBookBtn } from 'components/OrderBookBtn'
 
 // TradeWidget: subcomponents
 import { TradeFormData } from 'components/TradeWidget'
-import { FormInputError } from 'components/atoms/FormInputError'
-import { FormMessage } from 'components/atoms/FormMessage'
+import { FormInputError } from 'components/common/FormInputError'
+import { FormMessage } from 'components/common/FormMessage'
 import { useNumberInput } from 'components/TradeWidget/useNumberInput'
 import { SwapPrice } from './PriceEstimations'
