+++ src/components/TradeWidget/TokenRow.tsx
@@ -14,13 +14,13 @@ import Form from 'components/DepositWidget/Form'
 import { TooltipWrapper, HelpTooltipContainer, HelpTooltip } from 'components/Tooltip'
 import { InputBox } from 'components/InputBox'
 import { Input } from 'components/Input'
-import { Spinner } from 'components/atoms/Spinner'
+import { Spinner } from 'components/common/Spinner'
 import { WrapEtherBtn } from 'components/WrapEtherBtn'
 
 // TradeWidget: subcomponents
 import { TradeFormTokenId, TradeFormData } from 'components/TradeWidget'
-import { FormInputError } from 'components/atoms/FormInputError'
-import { FormMessage } from 'components/atoms/FormMessage'
+import { FormInputError } from 'components/common/FormInputError'
+import { FormMessage } from 'components/common/FormMessage'
 import { faPlus } from '@fortawesome/free-solid-svg-icons'
 
 // Hooks
