+++ src/components/common/FormInputError.stories.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { FormInputError, Props } from 'components/atoms/FormInputError'
+import { FormInputError, Props } from 'components/common/FormInputError'
 
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta, Story } from '@storybook/react/types-6-0'
 
 export default {
-  title: 'Atoms/FormInputError',
+  title: 'Common/FormInputError',
 } as Meta
 
 const Template: Story<Props> = (args) => <FormInputError {...args} />
